Fixed bugs in the clan boss battle simulator: 1. The turn meter calculations were based on floats but needed to be doubles. 2. The turn meter magic number was based on the fraction but needed to be based on the decimal value. 3. The tiebreaker for champions with the same speed needed to be their order in the battle, not arbitrary.
Add initial support for searching for unkillable teams as deltas of existing teams.  Let champions create clones of themselves with deltas in uiSpeed and a new speed set count.

Add unit test library to test clan boss battle simulator.

+++ RaidLib/Simulator/ChampionInBattle.cs
@@ -105,9 +105,9 @@ namespace RaidLib.Simulator
 
         public Champion Champ { get; private set; }
 
-        public float TurnMeter { get; private set; }
+        public double TurnMeter { get; private set; }
 
-        public float TurnMeterIncreaseOnClockTick { get; private set; }
+        public double TurnMeterIncreaseOnClockTick { get; private set; }
 
         public Skill TakeTurn()
         {
